+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - The dragon king Suzaku has been prepared for battles (prototype assets).
 - The layout for a debug menu for custom battles has been added.
 - A to-do list has been added for programmers to keep track of what we still need.
+- The menu for custom battles would write battle data onto a text file, but the battle scene won't receive that information as of yet.